Title: 체질량지수, Time: 121ms, Memory: 10MB, Status: Passed - CodeTree

+++ 231205/체질량지수/bmi.java
@@ -12,7 +12,7 @@ public class Main {
             System.out.println((int)bmi);
             System.out.print("Obesity");
         } else {
-            System.out.print(bmi);
+            System.out.print((int)bmi);
         }
     }
 }
